// services/admin-ui/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        OrganizationCreate: {
            /** Id */
            id: string;
            /** Name */
            name: string;
            /** Org Type */
            org_type?: string | null;
            /** Branch Id */
            branch_id?: string | null;
            /** Sub District */
            sub_district?: string | null;
            /** District */
            district?: string | null;
            /** Province */
            province?: string | null;
            /** Email */
            email?: string | null;
            /** Max Participants */
            max_participants?: number | null;
            /**
             * Gender Male
             * @default 0
             */
            gender_male: number;
            /**
             * Gender Female
             * @default 0
             */
            gender_female: number;
            /**
             * Gender Unspecified
             * @default 0
             */
            gender_unspecified: number;
            /** Contact Name */
            contact_name?: string | null;
            /** Contact Phone */
            contact_phone?: string | null;
            /** Contact Line Id */
            contact_line_id?: string | null;
            /** Enrolled Date */
            enrolled_date?: string | null;
            /** Enrolled Until */
            enrolled_until?: string | null;
            /** Latitude */
            latitude?: number | null;
            /** Longitude */
            longitude?: number | null;
            /** Contact */
            contact?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}